requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.2
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Flask==2.2.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
openslide-python==1.2.0
Pillow==9.4.0
pytz==2023.3
sqlparse==0.4.4
Werkzeug==2.2.3
zipp==3.15.0
